@@ -2,7 +2,7 @@ import { Router } from 'express';
 var router = Router();
 
 /* GET home page. */
-router.get('/', function(req, res, next) {
+router.get('/', function(req, res) {
   res.render('index', { title: "Welcome to the Fibonacci Calculator" });
 });
 
